+++ src/agent/sentient_agent_framework/implementation/default_response_handler.py
@@ -3,11 +3,11 @@ import inspect
 import json
 from cuid2 import Cuid
 from functools import wraps
-from src.agent.sentient_agent.interface.exceptions import (
+from src.agent.sentient_agent_framework.interface.exceptions import (
     AgentError,
     ResponseStreamClosedError
 )
-from src.agent.sentient_agent.interface.events import (
+from src.agent.sentient_agent_framework.interface.events import (
     DocumentEvent,
     DoneEvent,
     ErrorContent,
@@ -16,10 +16,10 @@ from src.agent.sentient_agent.interface.events import (
     TextBlockEvent,
     DEFAULT_ERROR_CODE
 )
-from src.agent.sentient_agent.interface.hook import Hook
-from src.agent.sentient_agent.interface.identity import Identity
-from src.agent.sentient_agent.interface.stream_event_emitter import StreamEventEmitter
-from src.agent.sentient_agent.implementation.default_text_stream import DefaultTextStream
+from src.agent.sentient_agent_framework.interface.hook import Hook
+from src.agent.sentient_agent_framework.interface.identity import Identity
+from src.agent.sentient_agent_framework.interface.stream_event_emitter import StreamEventEmitter
+from src.agent.sentient_agent_framework.implementation.default_text_stream import DefaultTextStream
 from typing import (
     Any,
     Callable,
