+++ src/agent/sentient_agent_framework/interface/exceptions.py
