+++ src/agent/sentient_agent_framework/interface/stream_event_emitter.py
